move createUser to UserRepository, remove methods from dao, delete dao

+++ HomeWork/HomeWork_5/notebook/view/UserView.java
@@ -4,7 +4,10 @@ import controller.UserController;
 import model.User;
 import util.Commands;
 
+import java.util.List;
 import java.util.Scanner;
+
+import static model.User.getDataUser;
 
 public class UserView {
     private final UserController userController;
@@ -17,16 +20,19 @@ public class UserView {
         Commands com;
 
         while (true) {
-            String command = prompt("Введите команду: ");
+            String command = prompt("Enter command: ");
             com = Commands.valueOf(command.toUpperCase()); // убрал зависимость от регистра
             if (com == Commands.EXIT) return;
             switch (com) {
                 case CREATE:
-                    User u = createUser();
-                    userController.saveUser(u);
+                    List<String> dataNewUser = getDataUser();
+                    userController.saveUser(userController.createUser(dataNewUser));
+                    System.out.println();
+                    break;
+                case NONE:
                     break;
                 case READ:
-                    String id = prompt("Идентификатор пользователя: ");
+                    String id = prompt("Enter user id: ");
                     try {
                         User user = userController.readUser(Long.parseLong(id));
                         System.out.println(user);
@@ -37,19 +43,31 @@ public class UserView {
                     break;
                 case DELETE:
                     String uId = prompt("Enter user id: ");
+                    if (uId.isEmpty()) {
+                        throw new RuntimeException("ID can not be null");
+                    }
                     userController.deleteUser(uId);
+                    System.out.println();
                     break;
                 case UPDATE:
                     String userId = prompt("Enter user id: ");
-                    userController.updateUser(userId, createUser());
+                    if (userId.isEmpty()) {
+                        throw new RuntimeException("ID can not be null");
+                    }
+                    List<String> dataCreateUser = getDataUser();
+                    userController.updateUser(userId, userController.createUser(dataCreateUser));
+                    System.out.println();
                     break;
                 case LIST:
                     System.out.println(userController.getAllUsers());
+                    System.out.println();
 
 
             }
         }
     }
+
+
 
     private String prompt(String message) {
         Scanner in = new Scanner(System.in);
@@ -57,20 +75,20 @@ public class UserView {
         return in.nextLine();
     }
 
-    private User createUser() {
-        String firstName = prompt("Имя: ");
-        if (firstName.isEmpty()) {
-            throw new RuntimeException("Имя не может быть пустым");
-        }
-        String lastName = prompt("Фамилия: ");
-        if (lastName.isEmpty()) {
-            throw new RuntimeException("Фамилия не может быть пустой");
-        }
-        String phone = prompt("Номер телефона: ");
-        if (phone.isEmpty()) {
-            throw new RuntimeException("Телефон не может быть пустым");
-        }
-
-        return new User(firstName.replaceAll(" ", ""), lastName.replaceAll(" ", ""), phone);
-    }
+//    private User createUser() {
+//        String firstName = prompt("Имя: ");
+//        if (firstName.isEmpty()) {
+//            throw new RuntimeException("Имя не может быть пустым");
+//        }
+//        String lastName = prompt("Фамилия: ");
+//        if (lastName.isEmpty()) {
+//            throw new RuntimeException("Фамилия не может быть пустой");
+//        }
+//        String phone = prompt("Номер телефона: ");
+//        if (phone.isEmpty()) {
+//            throw new RuntimeException("Телефон не может быть пустым");
+//        }
+//
+//        return new User(firstName.replaceAll(" ", ""), lastName.replaceAll(" ", ""), phone.replaceAll(" ", ""));
+//    }
 }
